+++ src/main/java/io/hhplus/tdd/point/entity/UserPoint.java
@@ -1,11 +1,11 @@
-package io.hhplus.tdd.point;
+package io.hhplus.tdd.point.entity;
 
 public record UserPoint(
         long id,
         long point,
         long updateMillis
-) {
-
+    )
+{
     public static UserPoint empty(long id) {
         return new UserPoint(id, 0, System.currentTimeMillis());
     }
